+++ Assets/DesignData/MapData/MapSepcificData/BrawlModeReforgedModeData.asset
@@ -12,11 +12,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: d5ccb7593c2726a4187b9d10fa8ab032, type: 3}
   m_Name: BrawlModeReforgedModeData
   m_EditorClassIdentifier: 
-  CanonFireTime: 0
-  CanonFireAlertTime: 0
-  CanonCoolDown: 0
-  CanonRadius: 0
-  CanonPower: 0
-  BagelKillPoint: 0
-  NormalKillPoint: 0
-  TotalTime: 0
+  CanonFireTime: 4
+  CanonFireAlertTime: 1
+  CanonCooldown: 15
+  FollowSpeed: 3
+  CanonRadius: 2
+  CanonPower: 1200
+  BagelKillPoint: 2
+  NormalKillPoint: 1
+  TotalTime: 180
