+++ Assets/Scripts/Chang/Reforged/BrawlModeReforgedWeaponManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 196d596685aba674da713a86687073d8
+guid: e3b8bc86fc0fcbf4daa54edd2dfec4e8
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
