+++ Assets/Scripts/Events.cs
@@ -322,6 +322,11 @@ public class WeaponSpawned : GameEvent
     }
 }
 
+public class BagelDespawn: GameEvent
+{
+
+}
+
 public class ObjectPickedUp : GameEvent
 {
     public GameObject Player;
